pyproject.toml
[project]
name = "qrgenerator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=11.1.0",
    "qrcode>=8.0",
    "ruff>=0.9.10",
]
